refactor: break up the interfaces a bit
