+++ packages/react-dnd/src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './classApi'
+export * from './monitors'
+export * from './hooksApi'
+export * from './options'
